Move auth logic into separate module
