+++ docker-compose.yml
@@ -18,7 +18,11 @@ services:
     image: rabbitmq:latest
 
   db:
-    image: openstax/cnx-db:2.3.0
+    # image: openstax/cnx-db:2.3.0
+    build:
+      context: https://github.com/Connexions/cnx-db.git
+      args:
+        gitcommithash: 8cafdfa
     ports:
       - "5432:5432"
     environment:
