+++ press/config.py
@@ -2,38 +2,10 @@ import os.path
 import warnings
 
 from pyramid.config import Configurator
-from pyramid.authentication import BasicAuthAuthenticationPolicy
-from pyramid.authorization import ACLAuthorizationPolicy
 from pyramid.config.settings import asbool
 from sqlalchemy.exc import SAWarning
-from pyramid.security import Everyone, Authenticated, Allow
+from .auth import RootFactory
 
-
-import hashlib
-from base64 import decodestring as decode
-def check_password(pass_hash, password):
-    challenge_bytes = decode(pass_hash[6:])
-    digest = challenge_bytes[:20]
-    salt = challenge_bytes[20:]
-    hr = hashlib.sha1(password.encode('utf8'))
-    hr.update(salt)
-    return digest == hr.digest()
-
-class RootFactory(object):
-    """Application root object factory.
-    Everything is accessed from the root, so the acls defined here
-    are applied to all requests.
-    """
-
-    __acl__ = (
-        (Allow, Authenticated, 'view'),
-    )
-
-    def __init__(self, request):
-        self.request = request
-
-    def __getitem__(self, key):  # pragma: no cover
-        raise KeyError(key)
 
 def discover_set(settings, setting_name, env_var, default=None,
                  modifier=None):
@@ -63,13 +35,6 @@ def discover_set(settings, setting_name, env_var, default=None,
         settings.setdefault(setting_name, default)
 
 
-def check_credentials(username, password, request):
-    """Returns a sequence of principal identifiers for the user.
-    """
-    _hash = b'{SSHA}Tf5uQqRItW5v4j0WDM5w3cgIwfKKATpX'
-    if check_password(_hash, password):
-        return [username]
-
 def configure(settings=None):
     """Configure the :mod:`pyramid.configure.Configurator` object"""
     if settings is None:
@@ -93,9 +58,7 @@ def configure(settings=None):
     config.include('.subscribers')
     config.include('.views')
     config.include('.tasks')
-    auth_policy = BasicAuthAuthenticationPolicy(check_credentials)
-    config.set_authentication_policy(auth_policy)
-    config.set_authorization_policy(ACLAuthorizationPolicy())
+    config.include('.auth')
     with warnings.catch_warnings():
         warnings.simplefilter("ignore", category=SAWarning)
         config.include('cnxdb.contrib.pyramid')
