+++ press/views/legacy_publishing.py
@@ -17,10 +17,7 @@ from ..utils import (
     convert_version_tuple_to_version_string,
     convert_version_to_legacy_version
 )
-# Security
-from pyramid.httpexceptions import HTTPUnauthorized
-from pyramid.security import forget
-from pyramid.view import forbidden_view_config
+
 
 @view_config(route_name='api.v3.publications', request_method=['POST'],
              renderer='json', http_cache=0)
