+++ press/views/ping.py
@@ -8,8 +8,8 @@ def ping(request):
     """A ping and ack view for checking the service is up and running."""
     return Response('pong')
 
+
 @view_config(route_name='auth-ping', permission='view')
 def authedping(request):
     """A ping and ack view for checking the service is up and running."""
-    # import dpb; pdb.set_trace()
     return Response('pong')
